move AppInfoExt from android.ext.settings to android.ext package
