+++ services/core/java/com/android/server/ext/PackageManagerHooks.java
@@ -7,7 +7,7 @@ import android.app.AppBindArgs;
 import android.content.pm.GosPackageState;
 import android.content.pm.PackageManager;
 import android.content.pm.PackageManagerInternal;
-import android.ext.settings.AppInfoExt;
+import android.ext.AppInfoExt;
 import android.location.HookedLocationManager;
 import android.os.Binder;
 import android.os.Build;
